+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Max compression profile is now free for everyone.
+
 ## [1.11.4] - 2026-07-30
 
 ### Fixed